pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.7",
    "dhanhq>=2.0.2",
    "eventlet>=0.40.3",
    "flask-login>=0.6.3",
    "flask-socketio>=5.5.1",
    "flask-wtf>=1.2.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
